Refactors Theme
