+++ src/components/Theme/index.jsx
@@ -1,4 +1,2 @@
 export { default as Theme } from "./Theme";
-export { default as ThemeWrapper } from "./ThemeWrapper";
 export { default as GlobalStyle } from "./GlobalStyle";
-export { adjustThemeColors } from "./ThemeUtilities";
